+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable, and occasionally less notable changes to this project will be docum
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.7.1] - Split Type and Relation
+- now you can specify rel= and is/prim= at the item level, so the actual
+type (Ulid/Principal, not Key) is used in the struct
+
 ## [0.7.0] - Removed Entity Nesting
 - Just wasn't working out, adding more complexity.  Much simplified now
 and 10% faster